Finishing Video and Variables
Unlike in the video my variables are not showing up a different color
(purple). Hovering over them says that they are unused local var
however in the video the variables were not used as well so I am
confused as to what is wrong here but I still set up the file.