Finishing Up
The only thing I could not fully understand was how to put variables in
the same line so I consoled out everything separately. I tried it the
way I knew from c++ with “my height is” << heightFeet “feet” and it did
not work and just made the whole sentence show up as 0.